chore: dark mode by default
Sets backgrounds to #000 by default. Someone using light mode will barely be affected or notice a black box on screen. Someone using dark mode will be able to read and see the interface, while not being blinded by a massive block of white text on white background.

There might be a better way to do this, like omitting the `background` attribute altogether, but I figure default dark mode is better than forcing light mode onto dark mode users. Even though I have Firefox set to adjust colors for dark mode, the interface here is always white, presumably because the background is forced to a color so Firefox only adjusts the color of the text and not the background.